+++ sdk/keyvault/azkeys/constants_test.go
@@ -16,13 +16,13 @@ func TestToPtrMethods(t *testing.T) {
 	d := DeletionRecoveryLevelCustomizedRecoverable
 	require.Equal(t, d.ToPtr(), &d)
 
-	j := JSONWebKeyCurveNameP256
+	j := KeyCurveNameP256
 	require.Equal(t, j.ToPtr(), &j)
 
-	o := JSONWebKeyOperationDecrypt
+	o := KeyOperationDecrypt
 	require.Equal(t, o.ToPtr(), &o)
 
-	a := KeyEncryptionAlgorithmRSAAESKEYWRAP256
+	a := KeyExportEncryptionAlgorithmRSAAESKEYWRAP256
 	require.Equal(t, a.ToPtr(), &a)
 }