+++ sdk/keyvault/azkeys/models.go
@@ -74,8 +74,8 @@ func keyAttributesFromGenerated(i *generated.KeyAttributes) *KeyAttributes {
 	}
 }
 
-// KeyBundle - A KeyBundle consisting of a WebKey plus its attributes.
-type KeyBundle struct {
+// KeyVaultKey - A KeyVaultKey consisting of a WebKey plus its attributes.
+type KeyVaultKey struct {
 	// The key management attributes.
 	Attributes *KeyAttributes `json:"attributes,omitempty"`
 
@@ -95,7 +95,7 @@ type KeyBundle struct {
 // JSONWebKey - As of http://tools.ietf.org/html/draft-ietf-jose-json-web-key-18
 type JSONWebKey struct {
 	// Elliptic curve name. For valid values, see JsonWebKeyCurveName.
-	Crv *JSONWebKeyCurveName `json:"crv,omitempty"`
+	Crv *KeyCurveName `json:"crv,omitempty"`
 
 	// RSA private exponent, or the D component of an EC private key.
 	D []byte `json:"d,omitempty"`
@@ -148,7 +148,7 @@ func jsonWebKeyFromGenerated(i *generated.JSONWebKey) *JSONWebKey {
 	}
 
 	return &JSONWebKey{
-		Crv:     (*JSONWebKeyCurveName)(i.Crv),
+		Crv:     (*KeyCurveName)(i.Crv),
 		D:       i.D,
 		DP:      i.DP,
 		DQ:      i.DQ,
@@ -246,9 +246,9 @@ func keyItemFromGenerated(i *generated.KeyItem) *KeyItem {
 	}
 }
 
-// DeletedKeyBundle - A DeletedKeyBundle consisting of a WebKey plus its Attributes and deletion info
-type DeletedKeyBundle struct {
-	KeyBundle
+// DeletedKey - A DeletedKey consisting of a WebKey plus its Attributes and deletion info
+type DeletedKey struct {
+	KeyVaultKey
 	// The url of the recovery object, used to identify and recover the deleted key.
 	RecoveryID *string `json:"recoveryId,omitempty"`